Only log failure to set node-role label
This is not a critical error, as node-roles can easily be added
manually after the maintenance.

We will need to investigate a proper fix at some point, but this change
will allow the maintenance to complete, even if some nodes may not get
the nice visible node-role label that's used by kubectl.